Add a local on-stack FAT cache for next_cluster
For some operations, e.g. seeking in a file, this makes a huge
performance difference since we avoid (re)reading the FAT for every node
in the linked list.
For other operations that call next_cluster only once there are no
negative implications since the memory for the FAT block would have been
allocated on the stack anyways.